generalize repeated code and address small fixes
