+++ src/components/Target.js
@@ -1,10 +1,7 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import styled from '@emotion/styled';
 import ComponentFactory from './ComponentFactory';
 import Permalink from './Permalink';
-import { isBrowser } from '../utils/is-browser';
-import useCopyClipboard from '../hooks/useCopyClipboard';
 
 // Based on condition isValid, split array into two arrays: [[valid, invalid]]
 const partition = (array, isValid) => {
@@ -17,31 +14,15 @@ const partition = (array, isValid) => {
 };
 
 const DescriptionTerm = ({ children, html_id, ...rest }) => {
-  const [copied, setCopied] = useState(false);
-  const [headingNode, setHeadingNode] = useState(null);
-  const url = isBrowser ? window.location.href.split('#')[0] + '#' + html_id : '';
-
-  useCopyClipboard(copied, setCopied, headingNode, url);
-
-  const handleClick = (e) => {
-    setCopied(true);
-  };
-
   return (
     <dt>
       {children.map((child, j) => (
         <ComponentFactory key={j} {...rest} nodeData={child} />
       ))}
-      <Permalink copied={copied} setHeadingNode={setHeadingNode} id={html_id} handleClick={handleClick} />
-      <HeaderBuffer id={html_id}></HeaderBuffer>
+      <Permalink id={html_id} description="definition" />
     </dt>
   );
 };
-
-const HeaderBuffer = styled.div`
-  margin-top: -150px;
-  position: absolute;
-`;
 
 DescriptionTerm.propTypes = {
   children: PropTypes.arrayOf(PropTypes.object).isRequired,
